+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Released ...
 
 Summary: TODO
 
+- Moved `label-position-tolerance` from unsigned type to double
+
 - Default PNG encoding method when `png` is supplied is now `png8:m=h`, so paletted png using hextree color quantization (#2028)
   Use `png32` now for full color png. More details at https://github.com/mapnik/mapnik/wiki/Image-IO.